Initial pass at documentation, with Sagemaker code included
The docs/ directory contains a skeleton of the final documentation with
some sections filled in. README.md contains details for building. Aside
from that all documentation is ReStructured text. Only the html target
has been tested.

The Sagemaker (training) code has gone through a few iterations and is
close to release ready. Note that the Jupyter notebook is not included
here.